Rename config to config-types to avoid user confusion
