+++ src/integration-tests/collect-bond.test.ts
@@ -6,7 +6,7 @@ import {
   Signer,
 } from '@ajna-finance/sdk';
 import { MAINNET_CONFIG } from './test-config';
-import { configureAjna } from '../config';
+import { configureAjna } from '../config-types';
 import {
   getProvider,
   resetHardhat,
